Fixed dev-server & Fix JS creation of HTML
Fixed the dev-server not being able to compile due to mixing of javascript module imports and node require imports.

Fixed the JavaScript creation of the HTML. Didn't realize I needed to parse the responses themselves which is a big oopsies on my part. Figured it out and got all the AIP fetch & parse functions working. After that I had to figure out how to actually get the container to be made and appended to `document.body` which ended up being a couple of async functions being used together, overall not bad but took a second to figure out.